Create index app and move index code and assets to new app.
