+++ index/tests.py
@@ -1,0 +1,8 @@
+"""Index tests.
+
+Author: Jared Paubel
+Version: 0.1
+"""
+from django.test import TestCase
+
+# Create your tests here.
