+++ index/views.py
@@ -1,0 +1,33 @@
+"""Index view functions.
+
+Author: Jared Paubel
+Version: 0.1
+"""
+from django.shortcuts import render
+import random
+from django.http import HttpRequest, HttpResponse
+
+from index.models import Index
+from blog.models import PhotoGallery, Post
+
+
+def index_view(request: HttpRequest) -> HttpResponse:
+    """Function-based view that sends index data to template.
+
+    Args:
+        request (HttpRequest): Request to process.
+
+    Returns:
+        HttpResponse: Response to provide to template.
+    """
+    context = {
+        'index_data': Index.objects.first(),
+        'posts': Post.objects.filter(status=1).order_by('-updated_on')[:2],
+    }
+    if (len(PhotoGallery.objects.all()) > 0):
+        random_gallery = random.choice(PhotoGallery.objects.all())
+        context['carousel_gallery_name'] = random_gallery
+        context['carousel_photos'] = random.choice(
+            random_gallery.galleries.all()
+        )
+    return render(request, 'index.html', context)
